requirements.txt
openai>=1.0.0
faiss-cpu
streamlit
numpy
tqdm
unstructured
python-dotenv
unstructured[all-docs]
